Update suggested fee recipient tests
